+++ backend/config/main.php
@@ -52,7 +52,7 @@ return [
             return Yii::$app->get('backendUrlManager');
         },
         'user' => [
-            'identityClass' => 'shop\entities\User',
+            'identityClass' => 'shop\entities\User\User',
             'enableAutoLogin' => true,
             'identityCookie' => [
                 'name' => '_identity',
