+++ frontend/tests/functional/ResendVerificationEmailCest.php
@@ -73,7 +73,7 @@ class ResendVerificationEmailCest
     {
         $I->submitForm($this->formId, $this->formParams('test@mail.com'));
         $I->canSeeEmailIsSent();
-        $I->seeRecord('shop\entities\User', [
+        $I->seeRecord('shop\entities\User\User', [
             'email' => 'test@mail.com',
             'username' => 'test.test',
             'status' => \common\shop\User::STATUS_WAIT
