+++ frontend/tests/unit/forms/VerifyEmailFormTest.php
@@ -45,11 +45,11 @@ class VerifyEmailFormTest extends \Codeception\Test\Unit
     {
         $model = new VerifyEmailForm('4ch0qbfhvWwkcuWqjN8SWRq72SOw1KYT_1548675330');
         $user = $model->verifyEmail();
-        expect($user)->isInstanceOf('shop\entities\User');
+        expect($user)->isInstanceOf('shop\entities\User\User');
 
         expect($user->username)->equals('test.test');
         expect($user->email)->equals('test@mail.com');
-        expect($user->status)->equals(\shop\entities\User::STATUS_ACTIVE);
+        expect($user->status)->equals(\shop\entities\User\User::STATUS_ACTIVE);
         expect($user->validatePassword('Test1234'))->true();
     }
 }
