+++ shop/services/auth/AuthService.php
@@ -2,7 +2,7 @@
 
 namespace shop\services\auth;
 
-use shop\entities\User;
+use shop\entities\User\User;
 use shop\forms\auth\LoginForm;
 use shop\repositories\UserRepository;
