+++ shop/tests/unit/entities/User/RequestSignupTest.php
@@ -3,7 +3,7 @@
 namespace shop\tests\unit\entities\User;
 
 use Codeception\Test\Unit;
-use shop\entities\User;
+use shop\entities\User\User;
 
 class RequestSignupTest extends Unit
 {
