+++ shop/tests/unit/forms/PasswordResetRequestFormTest.php
@@ -4,7 +4,7 @@ namespace shop\tests\unit\forms;
 
 use shop\forms\PasswordResetRequestForm;
 use common\fixtures\UserFixture as UserFixture;
-use shop\entities\User;
+use shop\entities\User\User;
 
 class PasswordResetRequestFormTest extends \Codeception\Test\Unit
 {
